finish work with email

+++ routes/courses.js
@@ -3,39 +3,68 @@ const Course = require('../models/course');
 const router = Router();
 const auth = require('../middleware/auth');
 
-router.get('/', async (req, res, next) => {
-    const courses = await Course.find().populate('userId');
-    res.render('courses', {title : 'Courses page', isCourses : true, courses: courses});
+const isOwner = (course, req) => course.userId.toString() === req.user.id;
+
+router.get('/', async (req, res) => {
+    try {
+        const courses = await Course.find().populate('userId');
+        res.render('courses', { title : 'Courses page',
+                                isCourses : true,
+                                courses: courses,
+                                userId: req.user ? req.user._id.toString() : null
+        });
+    } catch (e) {
+        console.log(e)
+    }
 });
 
 router.get('/:id/edit', auth, async (req, res) => {
    if (!req.query.allow) {
        return res.redirect('/');
    };
-   const course = await Course.findById(req.params.id);
-   res.render('course-edit', {title: `Edit ${course.title}`, course: course})
+   try {
+       const course = await Course.findById(req.params.id);
+       if (!isOwner(course, req)) {
+           return   res.redirect('/courses');
+       }
+       res.render('course-edit', {title: `Edit ${course.title}`, course: course});
+   } catch (e) {
+       console.log(e)
+   }
 });
 
-router.get('/:id', async (req, res, next) => {
-    const course = await Course.findById(req.params.id);
-    res.render('course', {title: `Course: ${course.title}`, course: course});
-});
-
-router.post('/edit', auth, async (req, res) => {
-    const {id} = req.body;
-    delete req.body.id;
-    await Course.findByIdAndUpdate(id, req.body);
-    res.redirect('/courses');
-});
-
-router.post('/remove', auth, async (req, res) => {
+router.get('/:id', async (req, res) => {
     try {
-        await Course.deleteOne({_id: req.body.id})
-        res.redirect('/courses')
+        const course = await Course.findById(req.params.id);
+        res.render('course', {title: `Course: ${course.title}`, course: course});
     } catch (e) {
         console.log(e)
     }
 });
 
+router.post('/edit', auth, async (req, res) => {
+    try {
+        const {id} = req.body;
+        delete req.body.id;
+        const course = await Course.findById(id);
+        if (!isOwner(course, req)) {
+            return   res.redirect('/courses');
+        };
+        Object.assign(course, req.body);
+        await course.save();
+        res.redirect('/courses');
+    } catch (e) {
+        console.log(e);
+    }
+});
+
+router.post('/remove', auth, async (req, res) => {
+    try {
+        await Course.deleteOne({_id: req.body.id, userId: req.user.id});
+        res.redirect('/courses')
+    } catch (e) {
+        console.log(e)
+    }
+});
 
 module.exports = router;
